+++ deploy/ansible/mycat/vars/var_mycat.yaml
@@ -12,15 +12,4 @@ schemas:
 # - "information-schema"
 # - "performance-schema"
 # - "sys"
-
-
-
-
-
-   
-   
-
-
-
- 
-
+# ------------------------------------------------
