--- monitor/ansible/mysql/config_mysql_monitor.retry
@@ -1,1 +1,0 @@
-cstudio
