--- monitor/ansible/mysql/template/userparameter.conf
@@ -1,8 +1,0 @@
-#针对MySQL连接数的监控
-UserParameter=mtls.MaxConnections,/usr/local/python/bin/python3 /usr/local/mtlsclient/monitor.py --host=127.0.0.1 -umtls -pmtls0352 MaxConnections
-UserParameter=mtls.ThreadsConnected,/usr/local/python/bin/python3 /usr/local/mtlsclient/monitor.py --host=127.0.0.1 -umtls -pmtls0352 ThreadsConnected
-UserParameter=mtls.ThreadsCached,/usr/local/python/bin/python3 /usr/local/mtlsclient/monitor.py --host=127.0.0.1 -umtls -pmtls0352 ThreadsCached
-UserParameter=mtls.Connections,/usr/local/python/bin/python3 /usr/local/mtlsclient/monitor.py --host=127.0.0.1 -umtls -pmtls0352 Connections
-
-
-#
